updated navbar
